requirements.txt
python==3.9.12
pandas
torch==2.1.2
pytorch-lightning==2.0.7
numpy
sklearn
PIL
transformers
csv
os
tensorflow==2.16.2
keras==3.4.1
efficientnet_pytorch==0.7.1
flair==0.13.1
distro==1.9.0
groq==0.8.0
python-decouple==3.8
rouge==1.0.1
rouge-score==0.1.2
nltk==3.8.1
requests==2.32.3
BeautifulSoup4==4.12.3
datasets
evaluate
transformers
peft
streamlit
pyvis
causal-learn
git+https://github.com/tomsons22/trustworthyAI.git@c2e3535eba28cabf8e1309f55bffc3e4d4d78c00#subdirectory=gcastle
cdt